+++ packages/core-js-compat/src/data.js
@@ -1262,6 +1262,8 @@ const data = {
   },
   'esnext.array.last-item': {
   },
+  'esnext.array.unique-by': {
+  },
   'esnext.async-iterator.constructor': {
   },
   'esnext.async-iterator.as-indexed-pairs': {