+++ tests/pure/index.js
@@ -145,6 +145,7 @@ import './es.weak-set';
 QUnit.module('ESNext');
 import './esnext.array.filter-out';
 import './esnext.array.is-template-object';
+import './esnext.array.unique-by';
 import './esnext.async-iterator.constructor';
 import './esnext.async-iterator.as-indexed-pairs';
 import './esnext.async-iterator.drop';